Update to 8.0.1 (2428)

+++ TMessagesProj/src/main/java/org/telegram/messenger/GoogleVoiceClientService.java
@@ -34,7 +34,7 @@ public class GoogleVoiceClientService extends SearchActionVerificationClientServ
                 if (!TextUtils.isEmpty(text)) {
                     String contactUri = intent.getStringExtra("com.google.android.voicesearch.extra.RECIPIENT_CONTACT_URI");
                     String id = intent.getStringExtra("com.google.android.voicesearch.extra.RECIPIENT_CONTACT_CHAT_ID");
-                    int uid = Integer.parseInt(id);
+                    long uid = Long.parseLong(id);
                     TLRPC.User user = MessagesController.getInstance(currentAccount).getUser(uid);
                     if (user == null) {
                         user = MessagesStorage.getInstance(currentAccount).getUserSync(uid);
