+++ TMessagesProj/src/main/java/org/telegram/messenger/voip/VoIPService.java
@@ -483,7 +483,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 	public boolean mutedByAdmin() {
 		ChatObject.Call call = groupCall;
 		if (call != null) {
-			int selfId = getSelfId();
+			long selfId = getSelfId();
 			TLRPC.TL_groupCallParticipant participant = call.participants.get(selfId);
 			if (participant != null && !participant.can_self_unmute && participant.muted && !ChatObject.canManageCalls(chat)) {
 				return true;
@@ -666,14 +666,14 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 			throw new IllegalStateException("No account specified when starting VoIP service");
 		}
 		classGuid = ConnectionsManager.generateClassGuid();
-		int userID = intent.getIntExtra("user_id", 0);
-		int chatID = intent.getIntExtra("chat_id", 0);
+		long userID = intent.getLongExtra("user_id", 0);
+		long chatID = intent.getLongExtra("chat_id", 0);
 		createGroupCall = intent.getBooleanExtra("createGroupCall", false);
 		hasFewPeers = intent.getBooleanExtra("hasFewPeers", false);
 		joinHash = intent.getStringExtra("hash");
-		int peerChannelId = intent.getIntExtra("peerChannelId", 0);
-		int peerChatId = intent.getIntExtra("peerChatId", 0);
-		int peerUserId = intent.getIntExtra("peerUserId", 0);
+		long peerChannelId = intent.getLongExtra("peerChannelId", 0);
+		long peerChatId = intent.getLongExtra("peerChatId", 0);
+		long peerUserId = intent.getLongExtra("peerUserId", 0);
 		if (peerChatId != 0) {
 			groupCallPeer = new TLRPC.TL_inputPeerChat();
 			groupCallPeer.chat_id = peerChatId;
@@ -836,7 +836,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		createGroupInstance(CAPTURE_DEVICE_CAMERA, false);
 	}
 
-	public int getCallerId() {
+	public long getCallerId() {
 		if (user != null) {
 			return user.id;
 		} else {
@@ -1321,7 +1321,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		tgVoip[CAPTURE_DEVICE_CAMERA].onSignalingDataReceive(data.data);
 	}
 
-	public int getSelfId() {
+	public long getSelfId() {
 		if (groupCallPeer == null) {
 			return UserConfig.getInstance(currentAccount).clientUserId;
 		}
@@ -1338,7 +1338,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		if (chat == null || groupCall == null || groupCall.call.id != update.call.id) {
 			return;
 		}
-		int selfId = getSelfId();
+		long selfId = getSelfId();
 		for (int a = 0, N = update.participants.size(); a < N; a++) {
 			TLRPC.TL_groupCallParticipant participant = update.participants.get(a);
 			if (participant.left) {
@@ -1748,7 +1748,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 				if (response != null) {
 					AndroidUtilities.runOnUIThread(() -> mySource[CAPTURE_DEVICE_CAMERA] = ssrc);
 					TLRPC.Updates updates = (TLRPC.Updates) response;
-					int selfId = getSelfId();
+					long selfId = getSelfId();
 					for (int a = 0, N = updates.updates.size(); a < N; a++) {
 						TLRPC.Update update = updates.updates.get(a);
 						if (update instanceof TLRPC.TL_updateGroupCallParticipants) {
@@ -1817,7 +1817,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 				TLRPC.Updates updates = (TLRPC.Updates) response;
 				AndroidUtilities.runOnUIThread(() -> {
 					if (tgVoip[CAPTURE_DEVICE_SCREEN] != null) {
-						int selfId = getSelfId();
+						long selfId = getSelfId();
 						for (int a = 0, N = updates.updates.size(); a < N; a++) {
 							TLRPC.Update update = updates.updates.get(a);
 							if (update instanceof TLRPC.TL_updateGroupCallConnection) {
@@ -1964,7 +1964,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		if (groupCall == null || tgVoip[CAPTURE_DEVICE_CAMERA] == null) {
 			return;
 		}
-		int selfId = getSelfId();
+		long selfId = getSelfId();
 		ArrayList<RequestedParticipant> participants = null;
 		for (int a = 0, N = unknown.length; a < N; a++) {
 			TLRPC.TL_groupCallParticipant p = groupCall.participantsBySources.get(unknown[a]);
@@ -2849,7 +2849,7 @@ public class VoIPService extends Service implements SensorEventListener, AudioMa
 		startForeground(ID_ONGOING_CALL_NOTIFICATION, builder.getNotification());
 	}
 
-	private void startRingtoneAndVibration(int chatID) {
+	private void startRingtoneAndVibration(long chatID) {
 		SharedPreferences prefs = MessagesController.getNotificationsSettings(currentAccount);
 		AudioManager am = (AudioManager) getSystemService(AUDIO_SERVICE);
 		boolean needRing = am.getRingerMode() != AudioManager.RINGER_MODE_SILENT;
